fix(hub): Remove circular dep from session
```
yarn run v1.22.5
$ madge --circular src/index.ts
Processed 5 files (1s) (2 warnings)

✖ Found 1 circular dependency!

1) hub.ts > interfaces.ts > scope.ts > session.ts
```

To remove this circular dep, we extract the SessionFlusher
from session.ts into it's own file.
