+++ packages/hub/src/index.ts
@@ -1,5 +1,6 @@
 export { addGlobalEventProcessor, Scope } from './scope';
-export { Session, SessionFlusher } from './session';
+export { Session } from './session';
+export { SessionFlusher } from './sessionFlusher';
 export {
   // eslint-disable-next-line deprecation/deprecation
   getActiveDomain,
